# server/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "공모아 API 문서",
    "description": "공모아 서버 API관련 문서입니다.",
    "version": "1.0.0"
  },
  "host": "https://gongmoa.site/api",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/": {
      "get": {
        "tags": [
          "관리자/Users"
        ],
        "summary": "관리자 회원 검색 관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "type": "string"
          },
          {
            "name": "user_type",
            "in": "query",
            "type": "string"
          },
          {
            "name": "keyword",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/setStatus": {
      "post": {
        "tags": [
          "관리자/Users"
        ],
        "summary": "관리자 회원 상태변경 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/user/getUserDetail": {
      "get": {
        "tags": [
          "관리자/Users"
        ],
        "summary": "관리자 회원 상세 불러오기 로직",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/modifyPassword": {
      "post": {
        "tags": [
          "관리자/Users"
        ],
        "summary": "관리자 회원 비밀번호 변경 관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/setStatus": {
      "post": {
        "tags": [
          "관리자/Users"
        ],
        "summary": "관리자 회원 상태변경 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "value": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/getUserDetail": {
      "get": {
        "tags": [
          "관리자/Users"
        ],
        "summary": "관리자 회원 상세 불러오기 로직",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/modifyPassword": {
      "post": {
        "tags": [
          "관리자/Users"
        ],
        "summary": "관리자 회원 비밀번호 변경 관련 로직",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}